Readd code style config and move DriveTrain interfaces
